+++ tests/test_script_runner.py
@@ -18,6 +18,7 @@
 """Tests for ScriptRunner."""
 
 import re
+import subprocess
 from pathlib import Path
 
 import pytest
@@ -26,21 +27,14 @@ from pytest_executable.script_runner import ScriptExecutionError, ScriptRunner
 
 from . import ROOT_DATA_DIR
 
-SCRIPT_PATH = ROOT_DATA_DIR / "test_runner_fixture/runner.sh"
-
-
-def test_execution_with_setting(tmp_path):
-    """Test script execution with placeholder replaced."""
-    runner = ScriptRunner(SCRIPT_PATH, {"nproc": 100}, tmp_path)
-    runner.run()
-    _assertions(tmp_path, "echo 100", "100", "")
+DATA_DIR = ROOT_DATA_DIR / "runners"
 
 
 def test_error_with_missing_setting(tmp_path):
     """Test error when a placeholder cannot be replaced."""
-    error_msg = "in .*tests/data/test_runner_fixture/runner.sh: 'nproc' is undefined"
+    error_msg = "in .*tests/data/runners/nproc.sh: 'nproc' is undefined"
     with pytest.raises(ValueError, match=error_msg):
-        ScriptRunner(SCRIPT_PATH, {}, tmp_path)
+        ScriptRunner(DATA_DIR / "nproc.sh", {}, tmp_path)
 
 
 def test_error_with_unreadable_script(tmp_path):
@@ -50,26 +44,52 @@ def test_error_with_unreadable_script(tmp_path):
         ScriptRunner(Path("/bin/bash"), {}, tmp_path)
 
 
+def test_execution_with_setting(tmp_path):
+    """Test script execution with placeholder replaced."""
+    script_path = DATA_DIR / "nproc.sh"
+    runner = ScriptRunner(script_path, {"nproc": 100}, tmp_path)
+    runner.run()
+    _assertions(tmp_path / script_path.name, "echo 100", "100", "")
+
+
+def test_execution_with_timeout(tmp_path):
+    """Test script execution with timeout."""
+    # with enough time
+    script_path = DATA_DIR / "timeout.sh"
+    runner = ScriptRunner(script_path, {"timeout": "2s"}, tmp_path)
+    runner.run()
+
+    # without enough time
+    runner = ScriptRunner(script_path, {"timeout": "0.1s"}, tmp_path)
+    error_msg = (
+        r"Command '\['/usr/bin/env', 'bash', 'timeout\.sh'\]' timed out after "
+        ".* seconds"
+    )
+    with pytest.raises(subprocess.TimeoutExpired, match=error_msg):
+        runner.run()
+
+
 def test_execution_error(tmp_path):
     """Test error when the script execution fails."""
     error_msg = "execution failure, see the stdout and stderr files in /"
-    runner = ScriptRunner(SCRIPT_PATH.parent / "error/runner.sh", {}, tmp_path)
+    script_path = DATA_DIR / "error.sh"
+    runner = ScriptRunner(script_path, {}, tmp_path)
     with pytest.raises(ScriptExecutionError, match=error_msg):
         runner.run()
 
     _assertions(
-        tmp_path,
+        tmp_path / script_path.name,
         "ls non-existing-file",
         "",
         "ls: (?:cannot access )?'?non-existing-file'?: No such file or directory",
     )
 
 
-def _assertions(tmp_path, script, stdout, stderr_regex):
+def _assertions(runner_path, script, stdout, stderr_regex):
     # check the content of the script, stdout and stderr files
-    with open(tmp_path / "runner.sh") as file_:
+    with runner_path.open() as file_:
         assert file_.read().strip() == script
-    with open(tmp_path / "runner.sh.stdout") as file_:
+    with (runner_path.with_suffix(".sh.stdout")).open() as file_:
         assert file_.read().strip() == stdout
-    with open(tmp_path / "runner.sh.stderr") as file_:
+    with (runner_path.with_suffix(".sh.stderr")).open() as file_:
         assert re.match(stderr_regex, file_.read())