+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ Added
 ~~~~~
 - Testing on MacOS.
 - :option:`--exe-test-module` cli option for setting the default test module
+- Add timeout setting for the runner execution.
 
 0.4.0 - 2020-05-03
 ------------------